Terminado (creo)

+++ src/factura/view/Facturas.java
@@ -6,42 +6,51 @@ import java.util.Scanner;
 
 import control.Conexión;
 import factura.entity.Factura;
+import factura.entity.NoExisteFactura;
 import venta.entity.NoExisteVenta;
 import view.InputTypes;
 
 public class Facturas {
-	
+
 	private Conexión conexión;
 	private Scanner scanner;
+
 	public Facturas(Conexión conexión, Scanner scanner) {
-		super();
 		this.conexión = conexión;
 		this.scanner = scanner;
 	}
+
 	public void add() throws NoExisteVenta{
 		Factura factura = RegistroFactura.ingresar(scanner);
 		String sql = "Insert into Factura (numVenta, NIT, nombre, descripción) values(?,?,?,?)";
-			try {
-				conexión.consulta(sql);
-				conexión.getSentencia().setInt(1, factura.getNumVenta());
+		try {
+			conexión.consulta(sql);
+			conexión.getSentencia().setInt(1, factura.getNumVenta());
 			conexión.getSentencia().setInt(2, factura.getNit());
 			conexión.getSentencia().setString(3, factura.getNombre());
 			conexión.getSentencia().setString(4, factura.getDescripción());
 			conexión.modificacion();
-			} catch (SQLException e) {
-				throw new NoExisteVenta();
-			}
-			
+		} catch (SQLException e) {
+			throw new NoExisteVenta();
+		}
 	}
-	public void delete() throws  SQLException {
+
+	public void delete() throws  SQLException, NoExisteFactura {
 		int numVenta = InputTypes.readInt("Número de venta: ", scanner);
-		String sql = "delete from factura where numVenta = ?";
+		String sql1 = "select * from factura where numVenta = ?";
+		conexión.consulta(sql1);
+		conexión.getSentencia().setInt(1, numVenta);
+		ResultSet resultSet = conexión.resultado();
+		if (resultSet.next()) {
+			String sql = "delete from factura where numVenta = ?";
 			conexión.consulta(sql);
 			conexión.getSentencia().setInt(1, numVenta);
 			conexión.modificacion();
-
+		} else {
+			throw new NoExisteFactura();
+		}
 	}
-	
+
 	public void update() throws SQLException, NoExisteVenta {
 		ResultSet resultSet;
 		Factura factura=null;
@@ -59,36 +68,36 @@ public class Facturas {
 			descripción = resultSet.getString("descripción");
 			numVenta = resultSet.getInt("numVenta");
 			factura = new Factura(numVenta, nit, nombre, descripción);
+
+			System.out.println(factura);
+			Menú.menúModificar(scanner, factura);
+
+			sql = "update factura set NIT=?, nombre = ?, descripción = ?  where numVenta = ?";
+
+			conexión.consulta(sql);
+			conexión.getSentencia().setInt(4, factura.getNumVenta());
+			conexión.getSentencia().setInt(1, factura.getNit());
+			conexión.getSentencia().setString(2, factura.getNombre());
+			conexión.getSentencia().setString(3, factura.getDescripción());
+			conexión.modificacion();
 		} else {
 			throw new NoExisteVenta();
 		}
+	}	
 
-		System.out.println(factura);
-		Menú.menúModificar(scanner, factura);
-
-		sql = "update factura set NIT=?, nombre = ?, descripción = ?  where numVenta = ?";
-
-		conexión.consulta(sql);
-		conexión.getSentencia().setInt(4, factura.getNumVenta());
-		conexión.getSentencia().setInt(1, factura.getNit());
-		conexión.getSentencia().setString(2, factura.getNombre());
-		conexión.getSentencia().setString(3, factura.getDescripción());
-		conexión.modificacion();
-	}
-	public void list() throws NoExisteVenta {
+	public void list() throws NoExisteVenta, SQLException {
 		Factura factura;
 		String sql = "select * from factura ";
-		try {
-			conexión.consulta(sql);
-			ResultSet resultSet = conexión.resultado();
-		while (resultSet.next()) {
-			factura = new Factura(resultSet.getInt("numVenta"), resultSet.getInt("NIT"), resultSet.getString("nombre"), resultSet.getString("descripción"));
-			System.out.println(factura);
-		}
-		} catch (SQLException e) {
+		conexión.consulta(sql);
+		ResultSet resultSet = conexión.resultado();
+		if (resultSet.next()) {
+			resultSet.previous();
+			while (resultSet.next()) {
+				factura = new Factura(resultSet.getInt("numVenta"), resultSet.getInt("NIT"), resultSet.getString("nombre"), resultSet.getString("descripción"));
+				System.out.println(factura);
+			}
+		} else {
 			throw new NoExisteVenta();
 		}
-		
 	}
-
 }
